Disturbance and No Try
Implemented the initial Disturbance module. Supports fire and two kinds
of grazing.  Removed the monstrous DataNotFound exception.
Paramaterized some hard coded fallback values.  Made some changes to
loops so that 1. they use C++11 stuff and 2. code is more consistent

+++ librvs/Succession/SuccessionDriver.cpp
@@ -257,8 +257,8 @@ int RVS::Succession::SuccessionDriver::relativeTimeInStage(int sclass)
 
 int RVS::Succession::SuccessionDriver::relativeTimeInStageAdjuster(double cover, std::map<string, string> strVals, std::map<string, double> numVals)
 {
-	int startAge = numVals["startAge"];
-	int yearsPassed = ((int)cover / (int)numVals["max_cov"]) + startAge;
+	double startAge = numVals["startAge"];
+	int yearsPassed = int(cover / numVals["max_cov"]) + int(startAge);
 	return yearsPassed;
 }
